requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy-Utils==0.41.2
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.2
flask-marshmallow==1.2.1
marshmallow_sqlalchemy==1.0.0
webargs==8.2.0
dynaconf==3.2.6
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-sqlalchemy-mock==0.1.7
pytest-cov==5.0.0
Flask-RESTful==0.3.10